% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{class_predict}
\alias{class_predict}
\title{Create confusion matrix of the predictions vs reference class in the
test set.}
\usage{
class_predict(fit, test_df, test_class)
}
\arguments{
\item{fit}{a caret model}

\item{test_df}{A test dataframe}

\item{test_class}{A vector of the classes}
}
\value{
A confusion matrix
}
\description{
Create confusion matrix of the predictions vs reference class in the
test set.
}
