% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compare_selected_features_barplot}
\alias{compare_selected_features_barplot}
\title{plot the counts of the selected genes though samples colored by class}
\usage{
compare_selected_features_barplot(df, class)
}
\arguments{
\item{df}{A subset of the data with selected genes as columns}

\item{class}{A vector of the class}
}
\description{
plot the counts of the selected genes though samples colored by class
}
