% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge.R
\name{deg_voom}
\alias{deg_voom}
\title{Select diffentially expressed genes(DGEs) from a voom object}
\usage{
deg_voom(v, class, keepn = 10)
}
\arguments{
\item{v}{A voom object}

\item{class}{A vector of assigned class for each sample e.g. "disease" or
"control"}

\item{keepn}{Number of DEGs to return}
}
\value{
A vector of selected DEGs
}
\description{
Select diffentially expressed genes(DGEs) from a voom object
}
