% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedded_methods.R
\name{eliminate_glm}
\alias{eliminate_glm}
\title{Eliminate features with glmnet}
\usage{
eliminate_glm(x, y, keepn = 100, fitControl)
}
\arguments{
\item{x}{train dataframe}

\item{y}{train class}

\item{keepn}{Number of genes to keep}

\item{fitControl}{a configuration list for caret package}
}
\value{
a vector of selected genes
}
\description{
Eliminate features with glmnet
}
