% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_specific_feature_selection.R
\name{eliminate_RF}
\alias{eliminate_RF}
\title{Eliminate features with Random forest by sorting variable importance.}
\usage{
eliminate_RF(x, y, keepn = 100, fitControl)
}
\arguments{
\item{x}{train dataframe}

\item{y}{train class}

\item{keepn}{Number of genes to keep}

\item{fitControl}{a configuration list for caret package}
}
\value{
a vector of selected genes
}
\description{
Eliminate features with Random forest by sorting variable importance.
}
