% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_filters.R
\name{filter0}
\alias{filter0}
\title{Basic filter which just removes very low expressed genes or genes that are
not expressed at all}
\usage{
filter0(counts, keepn = 10000)
}
\arguments{
\item{counts}{A counts matrix}

\item{keepn}{Number of genes to keep}
}
\value{
A subset of the counts matrix with the selected genes
}
\description{
Basic filter which just removes very low expressed genes or genes that are
not expressed at all
}
