% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_filters.R
\name{filter_cpm}
\alias{filter_cpm}
\title{filter low expressed genes using cpm}
\usage{
filter_cpm(counts, cpm_limit = 1, keepn = 100)
}
\arguments{
\item{counts}{A counts matrix}

\item{cpm_limit}{A limit for cpm}

\item{keepn}{Number of genes to keep}
}
\value{
A new counts matrix with selected genes
}
\description{
filter low expressed genes using cpm
}
