% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_filters.R
\name{filter_linear_correlation}
\alias{filter_linear_correlation}
\title{filter genes by linear correlation with the class variable}
\usage{
filter_linear_correlation(mat, class, keepn = 2000)
}
\arguments{
\item{mat}{a train matrix}

\item{class}{a class vector}

\item{keepn}{Number of genes to keep}
}
\value{
A subset of the counts matrix with the selected genes
}
\description{
filter genes by linear correlation with the class variable
}
