% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{fit_svmRadial}
\alias{fit_svmRadial}
\title{fit a svmRadial}
\usage{
fit_svmRadial(x, y, fitControl)
}
\arguments{
\item{x}{A train data.frame}

\item{y}{A vector of train classes}

\item{fitControl}{A configuration variable for caret}
}
\description{
fit a svmRadial
}
