% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_downloader.R
\name{get_class}
\alias{get_class}
\title{Get class vector from the rse_gene}
\usage{
get_class(rse_gene)
}
\arguments{
\item{rse_gene}{an object downloaded from recount}
}
\value{
a character vector of the assigned class
}
\description{
Get class vector from the rse_gene
}
