% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_downloader.R
\name{load_data}
\alias{load_data}
\title{download an Rdata object from the recount2 project to be extended to
other public RNASeq datasets}
\usage{
load_data(id)
}
\arguments{
\item{id}{Id of the study to be downloaded and analysed}
}
\value{
rse_gene An RangedSummarizedExperiment object
}
\description{
download an Rdata object from the recount2 project to be extended to
other public RNASeq datasets
}
