% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge.R
\name{normalize_deseq2}
\alias{normalize_deseq2}
\title{Normalize counts matrix with DESeq2}
\usage{
normalize_deseq2(counts, class)
}
\arguments{
\item{counts}{A counts matrix}

\item{class}{A vector of assigned class for each sample e.g. "disease" or
"control"}
}
\value{
DESeqDataSet object
}
\description{
Normalize counts matrix with DESeq2
}
