% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge.R
\name{normalize_voom}
\alias{normalize_voom}
\title{Normalize counts matrix with voom}
\usage{
normalize_voom(counts)
}
\arguments{
\item{counts}{A counts matrix with genes in rows and samples in columns}
}
\value{
voom object
}
\description{
Normalize counts matrix with voom
}
