% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pca}
\alias{pca}
\title{Plot top four principle components in a counts matrix colored with a class
variable, useful to see how well a gene subset of a counts matrix can classify
the data}
\usage{
pca(counts, class)
}
\arguments{
\item{counts}{A counts matrix}

\item{class}{A vector of assigned class for each sample e.g. "disease" or
"control"}
}
\value{
A grid of ggplot2 plots
}
\description{
Plot top four principle components in a counts matrix colored with a class
variable, useful to see how well a gene subset of a counts matrix can classify
the data
}
