% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_multipe_rocs}
\alias{plot_multipe_rocs}
\title{plot ROC curves from all models}
\usage{
plot_multipe_rocs(model_title_list)
}
\arguments{
\item{model_title_list}{a list of caret fit models named
with the name of the model}

\item{A}{list of caret models}
}
\description{
plot ROC curves from all models
}
