% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidystuff.R
\name{tidyClass}
\alias{tidyClass}
\title{Clean a class vector by keeping only alphabetical letters
to avoid errors when we use the class as a y variable in
ml models}
\usage{
tidyClass(class)
}
\arguments{
\item{class}{A class vector}
}
\value{
vector A tidy version of the class vector
}
\description{
Clean a class vector by keeping only alphabetical letters
to avoid errors when we use the class as a y variable in
ml models
}
